ref(network): prepare for Consensus decoupling
